BlogProject/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.8
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
psycopg2==2.9.9
Pygments==2.17.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
